+++ poetry/console/commands/add.py
@@ -1,8 +1,8 @@
 from .init import InitCommand
-from .venv_command import VenvCommand
+from .env_command import EnvCommand
 
 
-class AddCommand(VenvCommand, InitCommand):
+class AddCommand(EnvCommand, InitCommand):
     """
     Add a new dependency to <comment>pyproject.toml</>.
 
@@ -125,7 +125,7 @@ If you do not specify a version constraint, poetry will choose a suitable one ba
 
         installer = Installer(
             self.output,
-            self.venv,
+            self.env,
             self.poetry.package,
             self.poetry.locker,
             self.poetry.pool,