+++ poetry/packages/project_package.py
@@ -1,4 +1,9 @@
+from poetry.semver import VersionRange
+from poetry.semver import parse_constraint
+from poetry.version.markers import parse_marker
+
 from .package import Package
+from .utils.utils import create_nested_marker
 
 
 class ProjectPackage(Package):
@@ -10,6 +15,9 @@ class ProjectPackage(Package):
         self.include = []
         self.exclude = []
 
+        if self._python_versions == "*":
+            self._python_constraint = parse_constraint("~2.7 || >=3.4")
+
     def is_root(self):
         return True
 
@@ -19,3 +27,19 @@ class ProjectPackage(Package):
         dependency.is_root = True
 
         return dependency
+
+    @property
+    def python_versions(self):
+        return self._python_versions
+
+    @python_versions.setter
+    def python_versions(self, value):
+        self._python_versions = value
+
+        if value == "*" or value == VersionRange():
+            value = "~2.7 || >=3.4"
+
+        self._python_constraint = parse_constraint(value)
+        self._python_marker = parse_marker(
+            create_nested_marker("python_version", self._python_constraint)
+        )