+++ tests/console/commands/test_add.py
@@ -197,7 +197,7 @@ Package operations: 2 installs, 0 updates, 0 removals
 Writing lock file
 
   - Installing pendulum (1.4.4)
-  - Installing demo (0.1.0)
+  - Installing demo (0.1.0 ../distributions/demo-0.1.0-py2.py3-none-any.whl)
 """
 
     assert tester.get_display(True) == expected
@@ -237,7 +237,7 @@ Package operations: 2 installs, 0 updates, 0 removals
 Writing lock file
 
   - Installing pendulum (1.4.4)
-  - Installing demo (0.1.0)
+  - Installing demo (0.1.0 ../distributions/demo-0.1.0.tar.gz)
 """
 
     assert tester.get_display(True) == expected